backend/requirements.txt
flask
flask-cors
python-dotenv
requests
supabase
apscheduler
